requirements.txt
Flask>=2.2
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
Flask-SocketIO>=5.3
python-dotenv>=1.0
psycopg2-binary>=2.9
passlib[bcrypt]>=1.7
PyJWT>=2.6
eventlet>=0.33
gunicorn>=20.1
Flask-Cors>=3.1
alembic>=1.10

# Testing
pytest>=7.0
pytest-cov>=4.0
pytest-flask>=1.2
alembic>=1.10
